+++ src/main/java/com/bewitchment/common/brew/SetehsWastesBrew.java
@@ -1,9 +1,11 @@
 package com.bewitchment.common.brew;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockRedSandstone;
-import net.minecraft.block.BlockSandStone;
-import net.minecraft.block.BlockStairs;
+import java.util.HashMap;
+import java.util.Map;
+
+import javax.annotation.Nullable;
+
+import net.minecraft.block.*;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.EntityLivingBase;
@@ -13,10 +15,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-
-import javax.annotation.Nullable;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * This class was created by Arekkuusu on 24/04/2017.
@@ -49,7 +47,7 @@ public class SetehsWastesBrew extends BlockHitBrew {
 	@SuppressWarnings("deprecation")
 	@Override
 	public void safeImpact(BlockPos pos, @Nullable EnumFacing side, World world, int amplifier) {
-		int box = 1 + (int) ((float) amplifier / 2F);
+		int box = 1 + (int) (amplifier / 2F);
 
 		BlockPos posI = pos.add(box, box, box);
 		BlockPos posF = pos.add(-box, -box, -box);
